+++ NEWS.md
@@ -1,5 +1,9 @@
 # openxlsx2 (development version)
 
+## New features
+
+* Add `hide_blanks` argument to `wb_set_row_heights()`. This allows hiding all blank rows in a worksheet. [1464](https://github.com/JanMarvin/openxlsx2/pull/1464)
+
 ## Fixes
 
 * It is now possible to save a newly created workbook as `.xlsm` and it will open in spreadsheet software. [1453](https://github.com/JanMarvin/openxlsx2/pull/1453)